fix: point cloud dirty check on reearth/core
