+++ src/core/engines/Cesium/Feature/Raster/hooks.ts
@@ -11,6 +11,7 @@ import { useEffect, useMemo, useRef } from "react";
 import { useCesium } from "resium";
 
 import type { ComputedFeature, ComputedLayer, Feature, PolygonAppearance } from "../../..";
+import { usePick } from "../hooks";
 import { attachTag, extractSimpleLayer, extractSimpleLayerData, generateIDWithMD5 } from "../utils";
 
 import { Props } from "./types";
@@ -202,15 +203,6 @@ export const useMVT = ({
   }, [layer]);
 
   useImageryProvider(imageryProvider, layer?.id);
-};
-
-export const usePick = <T extends object, U extends keyof T>(
-  o: T | undefined | null,
-  fields: readonly U[],
-): Pick<T, U> | undefined => {
-  const p = useMemo(() => (o ? pick(o, fields) : undefined), [o, fields]);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  return useMemo(() => p, [JSON.stringify(p)]);
 };
 
 const polygonAppearanceFields: (keyof PolygonAppearance)[] = [
